Add files via upload
The uploaded SeisHandler-0.2.0.tar.gz is the latest version of SeisHandler, a Python package designed for efficient organization and management of seismic data files. This release provides enhanced features and improved performance, simplifying the process of matching and organizing seismic files by their names. It is tailored for seismic research professionals and enthusiasts who require a streamlined and effective way to handle large sets of seismic data. This version ensures compatibility with modern Python environments and extends the package's utility with new functionalities.